admin-mit-backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
PyYAML==6.0.1
bcrypt==4.1.2
cryptography==41.0.8
paramiko==3.4.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
websocket-client==1.6.4
Jinja2==3.1.2
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
hypothesis==6.92.1
psutil==5.9.6
openpyxl==3.1.2
kubernetes>=21.7.0,<25.0.0